# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KobeAI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/quizzes/{quizId}/start:
    get:
      operationId: startQuiz
      tags: [quizzes]
      summary: Start a quiz
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quiz started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizStartResponse"
  /v1/quizzes/{quizId}/submit:
    post:
      operationId: submitQuiz
      tags: [quizzes]
      summary: Submit quiz answers
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmitRequest"
      responses:
        "200":
          description: Quiz submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizSubmitResponse"
...
components:
  schemas:
...
    QuizQuestion:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        options:
          type: array
          items:
            type: string
        points:
          type: number
      required:
        - id
        - text
        - options
        - points
    QuizStartResponse:
      type: object
      properties:
        attempt_id:
          type: string
        quiz_id:
          type: string
        title:
          type: string
        subject:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
        time_limit_minutes:
          type: number
        total_points:
          type: number
      required:
        - attempt_id
        - quiz_id
        - title
        - subject
        - questions
        - time_limit_minutes
        - total_points
    QuizAnswer:
      type: object
      properties:
        question_id:
          type: string
        answer:
          type: string
      required:
        - question_id
        - answer
    QuizSubmitRequest:
      type: object
      properties:
        attempt_id:
          type: string
        answers:
          type: array
          items:
            $ref: "#/components/schemas/QuizAnswer"
      required:
        - attempt_id
        - answers
    QuizSubmitResponse:
      type: object
      properties:
        score:
          type: number
        total_points:
          type: number
        points_earned:
          type: number
        correct_answers:
          type: number
        total_questions:
          type: number
        message:
          type: string
      required:
        - score
        - total_points
        - points_earned
        - correct_answers
        - total_questions
        - message
...